+++ WinLess/Forms/aboutForm.cs
@@ -26,34 +26,6 @@ namespace WinLess
             
             // return the ProductVersion without the last '.0'
             return version.Substring(0, version.Length - 2);
-        }
-
-        private void aboutForm_Shown(object sender, EventArgs e)
-        {
-            //Force paint by calling Application.DoEvents();
-            Application.DoEvents();
-
-            bool lessCompilerUpdateAvailable = LessCompiler.IsCompilerUpdateAvailable();
-            checkingForUpdatesLabel.Visible = false;
-
-            if(lessCompilerUpdateAvailable)
-            {
-                if (MessageBox.Show(string.Format("WinLess uses the official LESS compiler, less.js, to compile your LESS files.\n\nA new version of less.js is available. Do you want to update less.js from {0} to {1}?", LessCompiler.GetCurrentCompilerVersion(), LessCompiler.GetAvailableCompilerVersion()), "Update LESS compiler?", MessageBoxButtons.YesNo) == DialogResult.Yes)
-                {
-                    //Show updatingLabel and force repaint by calling Application.DoEvents();
-                    updatingLabel.Visible = true;
-                    Application.DoEvents();
-                    
-                    LessCompiler.UpdateCompiler();
-
-                    Version newVersion = LessCompiler.GetCurrentCompilerVersion();
-                    lessjsVersionLabel.Text = newVersion.ToString();
-
-                    updatingLabel.Visible = false;
-
-                    MessageBox.Show(string.Format("Succesfully updated less.js to version {0}", newVersion), "LESS compiler update");
-                }
-            }            
-        }        
+        }       
     }
 }