+++ WinLess/Forms/settingsForm.cs
@@ -20,7 +20,7 @@ namespace WinLess
         {
             startWithWindowsCheckBox.Checked = Program.Settings.StartWithWindows;
             startMinimizedCheckBox.Checked = Program.Settings.StartMinified;
-            checkForLessUpdatesCheckbox.Checked = Program.Settings.CheckForLessUpdates;
+            useGloballyInstalledLessCheckbox.Checked = Program.Settings.UseGloballyInstalledLess;
             defaultMinifyCheckBox.Checked = Program.Settings.DefaultMinify;
             compileOnSaveCheckBox.Checked = Program.Settings.CompileOnSave;
             showSuccessMessagesCheckbox.Checked = Program.Settings.ShowSuccessMessages;
@@ -30,7 +30,7 @@ namespace WinLess
         {
             Program.Settings.StartWithWindows = startWithWindowsCheckBox.Checked;
             Program.Settings.StartMinified = startMinimizedCheckBox.Checked;
-            Program.Settings.CheckForLessUpdates = checkForLessUpdatesCheckbox.Checked;
+            Program.Settings.UseGloballyInstalledLess = useGloballyInstalledLessCheckbox.Checked;
             Program.Settings.DefaultMinify = defaultMinifyCheckBox.Checked;
             Program.Settings.CompileOnSave = compileOnSaveCheckBox.Checked;
             Program.Settings.ShowSuccessMessages = showSuccessMessagesCheckbox.Checked;