+++ WinLess/LessCompiler.cs
@@ -32,26 +32,6 @@ namespace WinLess
             return GetVersionFromCommandResult(result);
         }
 
-        public static Version GetAvailableCompilerVersion()
-        {            
-            CommandResult result = ExecuteNodePackageManagerCommand("view less version");
-
-            return GetVersionFromCommandResult(result);
-        }
-
-        public static bool IsCompilerUpdateAvailable()
-        {
-            Version currentVersion = GetCurrentCompilerVersion();
-            Version availableVersion = GetAvailableCompilerVersion();
-
-            return (currentVersion.CompareTo(availableVersion) < 0);
-        }
-
-        public static void UpdateCompiler()
-        {
-            ExecuteNodePackageManagerCommand("update less", true);
-        }
-
         private static Version GetVersionFromCommandResult(CommandResult result)
         {
             if (result.IsSuccess)
@@ -65,23 +45,11 @@ namespace WinLess
             return null;
         }
 
-        private static CommandResult ExecuteNodePackageManagerCommand(string arguments, bool elevated = false)
-        {
-            string fileName = string.Format("{0}\\node_modules\\.bin\\npm.cmd", Application.StartupPath);
-            if (elevated)
-            {
-                ExecuteElevatedCommand(fileName, arguments);
-	            return null;
-            }
-            else
-            {
-				return ExecuteCommand(fileName, arguments);  
-            }
-        }
-
         private static CommandResult ExecuteLessCommand(string arguments)
         {
-            string fileName = string.Format("{0}\\node_modules\\.bin\\lessc.cmd", Application.StartupPath);
+            string fileName = Program.Settings.UseGloballyInstalledLess
+                ? string.Format("{0}\\lessc.cmd", Application.StartupPath) 
+                : string.Format("{0}\\node_modules\\.bin\\lessc.cmd", Application.StartupPath);
             return ExecuteCommand(fileName, arguments);
         }
 
@@ -100,36 +68,11 @@ namespace WinLess
         string arguments = string.Format("\"{0}\" \"{1}\" --no-color", lessFile, cssFile);
             if (minify)
             {
-                arguments = string.Format("{0} --clean-css --clean-option=--selectors-merge-mode:ie8", arguments);
+                arguments = string.Format("{0} --clean-css=\"--compatibility=ie8 --advanced\"", arguments);
             }
 
             return arguments;
         }
-
-		private static void ExecuteElevatedCommand(string fileName, string arguments)
-		{
-			try
-			{
-				var process = new Process {
-					StartInfo = new ProcessStartInfo()
-					{
-						WorkingDirectory = Application.StartupPath,
-                        FileName = fileName,
-						Arguments = arguments,
-						UseShellExecute = true,
-                        Verb = "runas"
-					}
-				};
-
-				process.Start();
-				process.WaitForExit();
-			}
-
-			catch (Exception e)
-			{
-                ExceptionHandler.LogException(e);
-			}
-		}
 
         private static CommandResult ExecuteCommand(string fileName, string arguments){
             var result = new CommandResult();